start first C++/Fortran connection
The main fortran file has been reorganized in subroutines that are
called in the c++ program. The program is not usable since the
command line argument in not transfert to the fortran init function
yet. However, the cmake config is working (MPI package).

todo: transfer the CLI argument to fortran.